+++ src/ARma/pattern.cpp
@@ -49,7 +49,7 @@ namespace ARma {
 	{
 		return this->rotMat;
 	}
-	void Pattern::getExtrinsics(int patternSize, const Mat& cameraMatrix, const Mat& distortions)
+	void Pattern::getExtrinsics(double patternSize, const Mat& cameraMatrix, const Mat& distortions)
 	{
 
 		CvMat objectPts;//header for 3D points of pat3Dpts
@@ -71,14 +71,14 @@ namespace ARma {
 		pat3DPts[0].x = 0.0;
 		pat3DPts[0].y = 0.0;
 		pat3DPts[0].z = 0.0;
-		pat3DPts[1].x = patternSize;
+		pat3DPts[1].x = (float)patternSize;
 		pat3DPts[1].y = 0.0;
 		pat3DPts[1].z = 0.0;
-		pat3DPts[2].x = patternSize;
-		pat3DPts[2].y = patternSize;
+		pat3DPts[2].x = (float)patternSize;
+		pat3DPts[2].y = (float)patternSize;
 		pat3DPts[2].z = 0.0;
 		pat3DPts[3].x = 0.0;
-		pat3DPts[3].y = patternSize;
+		pat3DPts[3].y = (float)patternSize;
 		pat3DPts[3].z = 0.0;
 
 		cvInitMatHeader(&objectPts, 4, 3, CV_32FC1, pat3DPts);
